Changed things. fixed things, added things.

+++ src/main/java/com/shnupbups/allthethings/gui/GuiOven.java
@@ -77,7 +77,7 @@ public class GuiOven extends GuiContainer{
 	     this.drawTexturedModalRect(guiLeft + 7, guiTop + 10, 176, 0, 13, 64);
 	     
 	     if(tileentity.canOperate()) {
-	    	 this.drawTexturedModalRect(guiLeft + 81, guiTop + 39, 176, 64, tileentity.operateStatus * 24 / tileentity.getTimeNeeded(), 17);
+	    	 this.drawTexturedModalRect(guiLeft + 81, guiTop + 39, 176, 64, (int) (tileentity.operateStatus * 24 / tileentity.getTimeNeeded()), 17);
 	     }
 	}
